feat: BOOK_FORM_STEPS 상수 구조 변경 및 분리
- BOOK_FORM_STEPS를 form.ts 파일로 이동하여, 각 단계별 사용해야하는 필드들을 명확하게 정의했습니다.
